move s3/file-bucket functions into dedicated module
