+++ cloudigrade/util/aws/__init__.py
@@ -1,6 +1,5 @@
 """Helper utility module to wrap up common AWS operations."""
 import enum
-import gzip
 import json
 import logging
 import re
@@ -8,11 +7,11 @@ from functools import wraps
 
 import boto3
 from botocore.exceptions import ClientError
-from django.conf import settings
 from django.utils.translation import gettext as _
 
 from util.aws.sqs import delete_message_from_queue, extract_sqs_message, \
     receive_message_from_queue
+from util.aws.s3 import get_object_content_from_s3
 from util.exceptions import (AwsSnapshotCopyLimitError,
                              AwsSnapshotNotOwnedError,
                              AwsVolumeError,
@@ -518,35 +517,6 @@ def _verify_policy_action(session, action):
             return False
     except ClientError as e:
         return _handle_dry_run_response_exception(action, e)
-
-
-def get_object_content_from_s3(bucket, key, compression='gzip'):
-    """
-    Get the file contents from an S3 object.
-
-    Args:
-        bucket (str): The S3 bucket the object is stored in.
-        key (str): The S3 object key identified.
-        compression (str): The compression format for the stored file object.
-
-    Returns:
-        str: The string contents of the file object.
-
-    """
-    content = None
-    region = settings.S3_DEFAULT_REGION
-    s3_object = boto3.resource('s3', region_name=region).Object(bucket, key)
-
-    object_bytes = s3_object.get()['Body'].read()
-
-    if compression == 'gzip':
-        content = gzip.decompress(object_bytes).decode('utf-8')
-    elif compression is None:
-        content = object_bytes.decode('utf-8')
-    else:
-        logger.error(_('Unsupported compression format'))
-
-    return content
 
 
 def rewrap_aws_errors(original_function):
